Page Up and Page Down now move one screen width Adjust some types to reduce need for casting

+++ text-editor/src/main.rs
@@ -15,7 +15,7 @@ mod utils;
 
 #[derive(Clone, Copy, PartialEq)]
 pub struct WindowSize {
-    pub rows: u16,
+    pub rows: u32,
     pub columns: u16,
 }
